+++ repos/web/services/downloadFile.js
@@ -1,0 +1,15 @@
+export const downloadFile = (fileName) => (response) => {
+  // create file link in browser's memory
+  const href = URL.createObjectURL(response);
+
+  // create "a" HTML element with href to file & click
+  const link = document.createElement('a');
+  link.href = href;
+  link.setAttribute('download', fileName); //or any other extension
+  document.body.appendChild(link);
+  link.click();
+
+  // clean up "a" element & remove ObjectURL
+  document.body.removeChild(link);
+  URL.revokeObjectURL(href);
+}