+++ repos/core/src/pages/eis/changelog.md
@@ -28,6 +28,7 @@
   }}
 />
 ```
+- Added `apiCalls` property for UI-Render component and `onApplyPeriods` listener for `Button` component. If `apiCalls` contain `updateExperienceData` method, it will be calls on click button with `onApplyPeriods` listener
 
 ###  Fixes
 - Add TS annotation for  `getValidationErrors` listener